First iteration of pluggable schema registry
